# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Digital Marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  stock: number;
  category: string;
  /** @nullable */
  imageUrl: string | null;
  createdAt: string;
}

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  description: string;
  /** @minimum 0 */
  price: number;
  /** @minimum 0 */
  stock: number;
  /** @minLength 1 */
  category: string;
  imageUrl?: string;
}

export interface ProductUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  /** @minimum 0 */
  price?: number;
  /** @minimum 0 */
  stock?: number;
  category?: string;
  imageUrl?: string;
}

export interface PurchaseInput {
  /** @minimum 1 */
  quantity: number;
}

export interface MarketplaceStats {
  totalProducts: number;
  totalCategories: number;
  totalValue: number;
  inStockCount: number;
  outOfStockCount: number;
  lowStockCount: number;
}

export interface Category {
  name: string;
  productCount: number;
}

export interface ErrorResponse {
  error: string;
}

export type ListProductsParams = {
category?: string;
search?: string;
inStock?: boolean;
minPrice?: number;
maxPrice?: number;
sortBy?: ListProductsSortBy;
sortOrder?: ListProductsSortOrder;
};

export type ListProductsSortBy = typeof ListProductsSortBy[keyof typeof ListProductsSortBy];


export const ListProductsSortBy = {
  name: 'name',
  price: 'price',
  stock: 'stock',
  createdAt: 'createdAt',
} as const;

export type ListProductsSortOrder = typeof ListProductsSortOrder[keyof typeof ListProductsSortOrder];


export const ListProductsSortOrder = {
  asc: 'asc',
  desc: 'desc',
} as const;