# lib/api-zod/src/generated/types/marketplaceStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Digital Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface MarketplaceStats {
  totalProducts: number;
  totalCategories: number;
  totalValue: number;
  inStockCount: number;
  outOfStockCount: number;
  lowStockCount: number;
}