# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Digital Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  description: string;
  /** @minimum 0 */
  price: number;
  /** @minimum 0 */
  stock: number;
  /** @minLength 1 */
  category: string;
  imageUrl?: string;
}