# lib/api-zod/src/generated/types/purchaseInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Digital Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface PurchaseInput {
  /** @minimum 1 */
  quantity: number;
}